Changed Positioning system to gps
