+++ src/gps_demo.cpp
@@ -1,7 +1,4 @@
 #include <ros/ros.h>
-#include <std_msgs/Float32.h>
-#include <geometry_msgs/Twist.h>
-#include <tf/transform_listener.h>
 #include <projeto_semear/Pose.h>
 #include <projeto_semear/GetPose.h>
 #include <cstdint>
